refactor(dex): refactor to match best practice (#1018)
* feat: allow random parameter for vpool genesis on simulation

* chore: upadte changelog

* chore: refactor dex files

* chore: remove go work sum

* chore: add back go work sum

* chore: update changelog
