--- x/dex/types/message_exit_pool_test.go
@@ -1,41 +1,0 @@
-package types
-
-import (
-	"testing"
-
-	"github.com/NibiruChain/nibiru/x/testutil"
-
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-	"github.com/stretchr/testify/require"
-)
-
-func TestMsgExitPool_ValidateBasic(t *testing.T) {
-	tests := []struct {
-		name string
-		msg  MsgExitPool
-		err  error
-	}{
-		{
-			name: "invalid address",
-			msg: MsgExitPool{
-				Sender: "invalid_address",
-			},
-			err: sdkerrors.ErrInvalidAddress,
-		}, {
-			name: "valid address",
-			msg: MsgExitPool{
-				Sender: testutil.AccAddress().String(),
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := tt.msg.ValidateBasic()
-			if tt.err != nil {
-				require.ErrorIs(t, err, tt.err)
-				return
-			}
-			require.NoError(t, err)
-		})
-	}
-}
